PrefetchBuffer uploaded, Stallunit used nathan
So nathan's tb was used for my personal tb, which i did not upload. but the waveforms are generating a pulse stall flag, even when a stall is not supposed to be detected, which should not be happening.

Also prefetch buffer is outputting weird values, and I do not think it alternates between instruction slots (0, 1), it just outputs 0 slot until it hits a write protect, where it changes to 1 and then back to 0 slot again. This should be alternating constantly, and then re-copy data when a write protect is needed to be sure the instruction is not skipped.